+++ src/test/scala/silk/SmokeTest.scala
